backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Werkzeug==2.3.7
redis==5.0.1
requests==2.31.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
spacy==3.6.1
nltk==3.8.1
textblob==0.17.1
python-dotenv==1.0.0
cryptography==41.0.4
validators==0.22.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
selenium==4.15.0
webdriver-manager==4.0.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
celery==5.3.2
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
python-multipart==0.0.6
python-magic==0.4.27
exifread==3.0.0
twilio==8.10.0
google-api-python-client==2.103.0
transformers==4.34.0
torch==2.0.1
torchvision==0.15.2
elasticsearch==8.10.0
langdetect==1.0.9
polyglot==16.7.4
python-levenshtein==0.23.0
fuzzywuzzy==0.18.0
python-dateutil==2.8.2
pytz==2023.3
geopy==2.4.0
pycountry==22.3.13
imagehash==4.3.1
pytesseract==0.3.10
face-recognition==1.3.0
deepface==0.0.79
python-docx==0.8.11
PyPDF2==3.0.1
python-pptx==0.6.21
xlrd==2.0.1
openpyxl==3.1.2
python-telegram-bot==20.6
discord.py==2.3.2
slack-sdk==3.23.0
facebook-sdk==3.1.0
tweepy==4.14.0
instaloader==4.9.6
youtube-dl==2021.12.17
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-cloud-language==2.12.0
google-cloud-translate==3.12.1
google-cloud-automl==2.11.4
firebase-admin==6.2.0
boto3==1.28.62
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-textanalytics==5.3.0
ibm-watson==7.0.1
anthropic==0.3.11
openai==0.28.1
huggingface-hub==0.17.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.15
pinecone-client==2.2.4
weaviate-client==3.25.1
qdrant-client==1.6.4
milvus==2.3.1
pgvector==0.2.3
redis-om==0.2.1
mongodb==4.5.0
neo4j==5.13.0
cassandra-driver==3.28.0
influxdb==5.3.1
clickhouse-driver==0.2.6
apache-airflow==2.7.2
celery[redis]==5.3.2
flower==2.0.1
dramatiq==1.13.0
rq==1.15.1
schedule==1.2.0
APScheduler==3.10.4
croniter==1.4.1
python-crontab==3.0.0
supervisor==4.2.5
circus==0.18.0
uwsgi==2.0.22
gevent==23.7.0
eventlet==0.33.3
whitenoise==6.5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-silk==5.0.4
sentry-sdk[flask]==1.35.0
rollbar==0.16.3
bugsnag==4.7.1
newrelic==8.10.0
datadog==0.48.0
prometheus-client==0.17.1
statsd==4.0.1
structlog==23.1.0
colorlog==6.7.0
python-json-logger==2.0.7
loguru==0.7.2
elastic-apm==6.18.1
jaeger-client==4.8.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-exporter-prometheus==1.12.0rc1
